Project refactor in prep for rewrite
